chore: prepare Release 23 for release testing
