+++ meteor/server/migration/1_11_0.ts
@@ -1,0 +1,39 @@
+import { addMigrationSteps, CURRENT_SYSTEM_VERSION } from './databaseMigration'
+import { CoreSystem } from '../../lib/collections/CoreSystem'
+import { setExpectedVersion } from './lib'
+import { PeripheralDeviceAPI } from '../../lib/api/peripheralDevice'
+
+// Release 23
+addMigrationSteps('1.11.0', [
+	{
+		id: 'Fix serviceMessages in CoreSystem',
+		canBeRunAutomatically: true,
+		validate: () => {
+			const core = CoreSystem.findOne()
+			if (core) {
+				for (let [key, message] of Object.entries(core.serviceMessages)) {
+					if (typeof message.timestamp === 'string') {
+						return true
+					}
+				}
+				return false
+			}
+			return false
+		},
+		migrate: () => {
+			const core = CoreSystem.findOne()
+			if (core) {
+				for (let [key, message] of Object.entries(core.serviceMessages)) {
+					if (typeof message.timestamp !== 'number') {
+						core.serviceMessages[key] = {
+							...message,
+							timestamp: new Date(message.timestamp).getTime(),
+						}
+					}
+				}
+			}
+		},
+	},
+	setExpectedVersion('expectedVersion.playoutDevice', PeripheralDeviceAPI.DeviceType.PLAYOUT, '_process', '^1.10.0'),
+	// setExpectedVersion('expectedVersion.mosDevice', PeripheralDeviceAPI.DeviceType.MOS, '_process', '^1.4.3'),
+])
