+++ meteor/server/migration/databaseMigration.ts
@@ -68,9 +68,10 @@ import { storeSystemSnapshot } from '../api/snapshot'
  * 1.8.0: Release 20  (2019-05-12)
  * 1.9.0: Release 21  (never released)
  * 1.10.0: Release 22  (2020-08-17)
- * x.x.x: Release 23  (TBD)
+ * 1.11.0: Release 23  (TBD)
+ * x.x.x: Release 24  (TBD)
  */
-export const CURRENT_SYSTEM_VERSION = '1.10.0'
+export const CURRENT_SYSTEM_VERSION = '1.11.0'
 
 /**
  * These versions are not supported anymore (breaking changes occurred after these versions)
