+++ meteor/server/migration/migrations.ts
@@ -21,5 +21,6 @@ import './1_7_0'
 import './1_8_0'
 import './1_9_0'
 import './1_10_0'
+import './1_11_0'
 
 import './X_X_X'
